Moved theme settings to applications directory.
