+++ src/applications/theme/ThemeSettingColorArticle.svelte
